+++ paq8pxd.cpp
@@ -126,11 +126,13 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/stationarymap.hpp"
 #include "prt/mft.hpp"
 #include "prt/contextmap2.hpp"
+#include "prt/contextmap.hpp"
 
 #include "models/model.hpp"
 #include "models/sparsematch.hpp"
 #include "models/exe.hpp"
 #include "models/dec.hpp"
+#include "models/indirect.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -247,6 +249,8 @@ const U8 WRT_wrd1[256]={
 const U32 tri[4]={0,4,3,7}, trj[4]={0,6,6,12};
 
 int dt[1024];  // i -> 16K/(i+i+3)
+
+int n0n1[256];
 #define PNGFlag (1<<31)
 #define GrayFlag (1<<30)
 
@@ -367,288 +371,6 @@ public:
     return cp[0]!=0;
   }
 };
-
-// Context map for large contexts.  Most modeling uses this type of context
-// map.  It includes a built in RunContextMap to predict the last byte seen
-// in the same context, and also bit-level contexts that map to a bit
-// history state.
-//
-// Bit histories are stored in a hash table.  The table is organized into
-// 64-byte buckets alinged on cache page boundaries.  Each bucket contains
-// a hash chain of 7 elements, plus a 2 element queue (packed into 1 byte)
-// of the last 2 elements accessed for LRU replacement.  Each element has
-// a 2 byte checksum for detecting collisions, and an array of 7 bit history
-// states indexed by the last 0 to 2 bits of context.  The buckets are indexed
-// by a context ending after 0, 2, or 5 bits of the current byte.  Thus, each
-// byte modeled results in 3 main memory accesses per context, with all other
-// accesses to cache.
-//
-// On bits 0, 2 and 5, the context is updated and a new bucket is selected.
-// The most recently accessed element is tried first, by comparing the
-// 16 bit checksum, then the 7 elements are searched linearly.  If no match
-// is found, then the element with the lowest priority among the 5 elements
-// not in the LRU queue is replaced.  After a replacement, the queue is
-// emptied (so that consecutive misses favor a LFU replacement policy).
-// In all cases, the found/replaced element is put in the front of the queue.
-//
-// The priority is the state number of the first element (the one with 0
-// additional bits of context).  The states are sorted by increasing n0+n1
-// (number of bits seen), implementing a LFU replacement policy.
-//
-// When the context ends on a byte boundary (bit 0), only 3 of the 7 bit
-// history states are used.  The remaining 4 bytes implement a run model
-// as follows: <count:7,d:1> <b1> <unused> <unused> where <b1> is the last byte
-// seen, possibly repeated.  <count:7,d:1> is a 7 bit count and a 1 bit
-// flag (represented by count * 2 + d).  If d=0 then <count> = 1..127 is the
-// number of repeats of <b1> and no other bytes have been seen.  If d is 1 then
-// other byte values have been seen in this context prior to the last <count>
-// copies of <b1>.
-//
-// As an optimization, the last two hash elements of each byte (representing
-// contexts with 2-7 bits) are not updated until a context is seen for
-// a second time.  This is indicated by <count,d> = <1,0> (2).  After update,
-// <count,d> is updated to <2,0> or <1,1> (4 or 3).
-#if defined(__AVX2__)
-#define MALIGN 32
-#else
-#define MALIGN 16
-#endif
-
-int n0n1[256];
-class ContextMap {
-  const int C;  // max number of contexts
-  class E {  // hash element, 64 bytes
-    U16 chk[7];  // byte context checksums
-    U8 last;     // last 2 accesses (0-6) in low, high nibble
-  public:
-    U8 bh[7][7]; // byte context, 3-bit context -> bit history state
-      // bh[][0] = 1st bit, bh[][1,2] = 2nd bit, bh[][3..6] = 3rd bit
-      // bh[][0] is also a replacement priority, 0 = empty
-    U8* get(U16 chk);  // Find element (0-6) matching checksum.
-      // If not found, insert or replace lowest priority (not last).
-  };
-  Array<E, 64> t;  // bit histories for bits 0-1, 2-4, 5-7
-    // For 0-1, also contains a run count in bh[][4] and value in bh[][5]
-    // and pending update count in bh[7]
-  Array<U8*> cp;   // C pointers to current bit history
-  Array<U8*> cp0;  // First element of 7 element array containing cp[i]
-  Array<U32> cxt;  // C whole byte contexts (hashes)
-  Array<U8*> runp; // C [0..3] = count, value, unused, unused
-  Array<short, MALIGN>  r0;   //for rle 
-  Array<short, MALIGN>  r1;
-  Array<short, MALIGN>  r0i;
-  Array<short, MALIGN>  rmask; // mask for skiped context
-  StateMap *sm;    // C maps of state -> p
-  int cn;          // Next context to set by set()
-  //void update(U32 cx, int c);  // train model that context cx predicts c
-  Random rnd;
-  int mix1(Mixer& m, int cc, int bp, int c1, int y1);
-    // mix() with global context passed as arguments to improve speed.
-  int model;  
-public:
-  ContextMap(U64 m, int c=1,int mod=-1);  // m = memory in bytes, a power of 2, C = c
-  ~ContextMap();
-  void set();   // set next whole byte context to cx
-  void set(U32 cx, int next=-1);   // set next whole byte context to cx
-  void set(U64 cx, int next=-1);   // set next whole byte context to cx
-    // if next is 0 then set order does not matter
-  //int mix(Mixer& m) { return mix1(m, m.x.c0, m.x.bpos, m.x.buf(1), m.x.y);}
-  int mix(Mixer& m) {return mix1(m, m.x.c0, m.x.bpos, m.x.buf(1), m.x.y);}
-    int inputs();
-};
-
-
-// Find or create hash element matching checksum ch
-inline U8* ContextMap::E::get(U16 ch) {
-    
-  if (chk[last&15]==ch) return &bh[last&15][0];
-  int b=0xffff, bi=0;
-  for (int i=0; i<7; ++i) {
-    if (chk[i]==ch) return last=last<<4|i, (U8*)&bh[i][0];
-    int pri=bh[i][0];
-    if (pri<b && (last&15)!=i && last>>4!=i) b=pri, bi=i;
-  }
-  return last=0xf0|bi, chk[bi]=ch, (U8*)memset(&bh[bi][0], 0, 7);
-}
-
-// Construct using m bytes of memory for c contexts(c+7)&-8
-ContextMap::ContextMap(U64 m, int c,int mod): C(c),  t(m>>6), cp(C), cp0(C),
-    cxt(C), runp(C), r0(C),r1(C),r0i(C),rmask(C),cn(0),model(mod) {
-  assert(m>=64 && (m&m-1)==0);  // power of 2?
-  assert(sizeof(E)==64);
-  sm=new StateMap[C];
-  for (int i=0; i<C; ++i) {
-    cp0[i]=cp[i]=&t[0].bh[0][0];
-    runp[i]=cp[i]+3;
-  }
- // #ifndef NDEBUG 
- // printf("ContextMap t %0.2f mbytes\n",(((t.size()*sizeof(E)) +0.0)/1024)/1024);
- // #endif
-}
-#define CMBADLIMIT 8*32767
-ContextMap::~ContextMap() {
-  delete[] sm;
-  
-}
-inline void ContextMap::set() {
-    set(U32(0));
-  
-}
-// Set the i'th context to cx
-inline void ContextMap::set(U64 cx, int next) {
-    set(finalize64(cx,32));
-  
-}
-// Set the i'th context to cx
-inline void ContextMap::set(U32 cx, int next) {
-    if (cn==C) cn=0; // model bypass FIXME
-  int i=cn++;
-  i&=next;
-  assert(i>=0 && i<C);
-  if (cx==0){ cxt[i]=0; rmask[i]=0;}
-  else{
-  cx=cx*987654323+i;  // permute (don't hash) cx to spread the distribution
-  cx=cx<<16|cx>>16;
-  cxt[i]=cx*123456791+i;
-  rmask[i]=-1;
-  }
-}
-// Predict to mixer m from bit history state s, using sm to map s to
-// a probability.
-
-inline int mix2(Mixer& m, int s, StateMap& sm) {
-  int p1=sm.p(s,m.x.y);
-  if (s==0){
-    m.add(0); m.add(0);m.add(0);m.add(0);m.add(48+16);
-    return 0;
-  }else{
-  //int n0=-!nex(s,2);
-  //int n1=-!nex(s,3);
-    int st=stretch(p1)>>2;
-    m.add(st);
-    p1>>=4;
-    int p0=255-p1;
-    if (m.x.rm1)  m.add(p1-p0); else m.add(0); // ??
-    //m.add(st*(n1-n0));
-    //m.add((p1&n0)-(p0&n1));
-    //m.add((p1&n1)-(p0&n0));
-    int n01=n0n1[s];
-    if (n01){
-        p0=(n01<2)?255:0;
-        m.add((p1)-(p0));
-        if (n0n1[s]==2){ // -1
-              m.add((-st));
-        }else{
-              m.add((st));
-        }
-     }else {
-        m.add(0);
-        m.add(0);
-     }
-     m.add(0);
-  return s>0;
-  }
-}
-
-// Update the model with bit y1, and predict next bit to mixer m.
-// Context: cc=c0, bp=bpos, c1=buf(1), y1=y.
-int ContextMap::mix1(Mixer& m, int cc, int bp, int c1, int y1) {
-  // Update model with y
-  int result=0;
-  for (int i=0; i<cn; ++i) {
-    if (cp[i]) {
-      assert(cp[i]>=&t[0].bh[0][0] && cp[i]<=&t[t.size()-1].bh[6][6]);
-      assert(((long long)(cp[i])&63)>=15);
-      int ns=nex(*cp[i], y1);
-      if (ns>=204 && rnd() << ((452-ns)>>3)) ns-=4;  // probabilistic increment
-      *cp[i]=ns;
-    }
-
-    // Update context pointers
-    if (m.x.bpos>1 && runp[i][0]==0)
-    {
-     cp[i]=0;
-    }
-    else
-    {
-     U16 chksum=cxt[i]>>16;
-     U64 tmask=t.size()-1;
-     switch(m.x.bpos)
-     {
-      case 1: case 3: case 6: cp[i]=cp0[i]+1+(cc&1); break;
-      case 4: case 7: cp[i]=cp0[i]+3+(cc&3); break;
-      case 2: case 5: cp0[i]=cp[i]=t[(cxt[i]+cc)&tmask].get(chksum); break;
-      default:
-      {
-       cp0[i]=cp[i]=t[(cxt[i]+cc)&tmask].get(chksum);
-       // Update pending bit histories for bits 2-7
-       if (cp0[i][3]==2) {
-         const int c=cp0[i][4]+256;
-         U8 *p=t[(cxt[i]+(c>>6))&tmask].get(chksum);
-         p[0]=1+((c>>5)&1);
-         p[1+((c>>5)&1)]=1+((c>>4)&1);
-         p[3+((c>>4)&3)]=1+((c>>3)&1);
-         p=t[(cxt[i]+(c>>3))&tmask].get(chksum);
-         p[0]=1+((c>>2)&1);
-         p[1+((c>>2)&1)]=1+((c>>1)&1);
-         p[3+((c>>1)&3)]=1+(c&1);
-         cp0[i][6]=0;
-       }
-       // Update run count of previous context
-       if (runp[i][0]==0)  // new context
-         runp[i][0]=2, runp[i][1]=c1;
-       else if (runp[i][1]!=c1)  // different byte in context
-         runp[i][0]=1, runp[i][1]=c1;
-       else if (runp[i][0]<254)  // same byte in context
-         runp[i][0]+=2;
-       else if (runp[i][0]==255)
-         runp[i][0]=128;
-       runp[i]=cp0[i]+3;
-      } break;
-     }
-    }
-   }
-  
-    
-    for (int i=0; i<cn; ++i) {
-       if(cxt[i]){
-            // predict from bit context
-            int s = 0;
-            if (cp[i]) s = *cp[i];
-            if (s>0) result++;
-            mix2(m, s, sm[i]);
-       }else{
-            for (int i=0; i<(inputs()-1); i++)
-                m.add(0);     
-       }
-   }
- 
-
-    // predict from last byte in context
-     
-     for (int i=0; i<cn; ++i) {
-         U8 a=runp[i][0];
-         U8 b=runp[i][1];
-         r0[i]=a;
-         r1[i]=b;
-         r0i[i]=ilog(a+1);
-     }
-
-       
-    for (int i=0; i<cn; ++i) {
-        if (rmask[i] && ((r1[i  ]+256)>>(8-bp)==cc)) {
-            m.add(((r1[i  ]>>(7-bp)&1)*2-1) *((r0i[i  ])<<(2+(~r0[i  ]&1)))); }
-        else   m.add(0);
-      }
-
-   
-  if (bp==7) cn=0;
-  return result;
-}
-
-int ContextMap::inputs() {
-    return 6;
-}
 
 
 
@@ -7244,148 +6966,7 @@ int p(Mixer& m,int info,int val2=0){
 
 
 
-//////////////////////////// indirectModel /////////////////////
 
-// The context is a byte string history that occurs within a
-// 1 or 2 byte context.
-class indirectModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  const int N;
-  const int mem;
-  ContextMap cm2;
-  ContextMap cm3;
-  ContextMap cm4;
-  ContextMap cm5;
-  ContextMap cmt;
-  ContextMap cm0;
-  ContextMap cma;
-  ContextMap cmc;
-  Array<U32> t1;
-  Array<U16> t2;
-  Array<U16> t3;
-  Array<U16> t4;
-  Array<U32> t5; // 256K
-  IndirectContext<U32> iCtx;
-  LargeIndirectContext<uint32_t> iCtxLarge{ 18,8 }; // 11MB // hashBits, inputBits
-  U32 chars4; 
-public:
-  indirectModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf),N(15+1+9-13),mem(0x1000000*(level>9?1:1)),
-
-  cm2((mem/4),4,M_INDIRECT),
-  cmt((mem/2),3,M_INDIRECT),
-    cm0((mem/2),3,M_INDIRECT), 
-    cma((mem/4),3,M_INDIRECT), 
-    cmc((mem/4),3,M_INDIRECT), 
-    cm3((mem/4),3,M_INDIRECT), 
-    cm4((mem/4),3,M_INDIRECT), 
-    cm5((mem/4),3,M_INDIRECT), 
-   t1(256),
-   t2(0x10000), t3(0x8000),t4(0x8000),t5(256*256),
-   iCtx{16,8 },
-   chars4(0){
-  }
-  int inputs() {return (4*3+13)*cm2.inputs()+2;}
-  int nets() {return 0;}
-  int netcount() {return 0;}
-  void setContexts(){
-      
-  }
-int p(Mixer& m,int val1=0,int val2=0){
-  int j=0;
-  if (!x.bpos) {
-    U32 d=x.c4&0xffff;
-    U8 c=d&255;
-    U32 d2=(x.buf(1)&31)+   32*(x.buf(2)&31)+   1024*(x.buf(3)&31);
-    U32 d3=(x.buf(1)>>3&31)+32*(x.buf(3)>>3&31)+1024*(x.buf(4)>>3&31);
-    
-    U32& r1=t1[d>>8];
-    r1=r1<<8|c;
-    U16& r2=t2[x.c4>>8&0xffff];
-    r2=r2<<8|c;
-    U16& r3=t3[(x.buf(2)&31)+32*(x.buf(3)&31)+1024*(x.buf(4)&31)];
-    r3=r3<<8|c;
-    U16& r4=t4[(x.buf(2)>>3&31)+32*(x.buf(4)>>3&31)+1024*(x.buf(5)>>3&31)];
-    r4=r4<<8|c;
-    const U32 t=c|t1[c]<<8;
-    const U32 t0=d|t2[d]<<16;
-    const U32 ta=d2|t3[d2]<<16;
-    const U32 tc=d3|t4[d3]<<16;
-    const U8 pc=tolower(U8(x.c4>>8));
-    iCtx+=(c=tolower(c)), iCtx=(pc<<8)|c;
-    
-    const U32 ctx0=iCtx(), mask=(U8(t1[c])==U8(t2[d]))|
-                               ((U8(t1[c])==U8(t3[d2]))<<1)|
-                               ((U8(t1[c])==U8(t4[d3]))<<2)|
-                               ((U8(t1[c])==U8(ctx0))<<3);
-    cmt.set(t);
-    cmt.set( (t&0xff00)| mask);
-    cmt.set(t&0xffff);
-    
-    cm0.set(t0);
-    cm0.set(t0&0xffffff);
-    cm0.set((t0<<9)|x.frstchar);
-    
-    cma.set(ta); 
-    cma.set(ta&0xffffff);
-    cma.set((ta<<9)|x.frstchar);
-    
-    cmc.set(tc);
-    cmc.set((tc<<9)|x.frstchar);
-    cmc.set(buf(1) + ((32 * x.tt) & 0x1FFFFF00));
-    
-     // context with 2 characters converted to lowercase (context table: t5, byte history: h5)
-    U32 c1=c;
-    c = tolower(c1);
-    chars4 = chars4 << 8 | c;
-
-    const U32 h5 = t5[chars4 & 0xffff];
-    U32& r5 = t5[(chars4 >> 8) & 0xffff];
-    r5 = r5 << 8 | c;
-
-    cm2.set( (h5 & 0xff) | (c << 8)); //last char + 1 history char
-    cm2.set(  h5 & 0xffff); //2 history chars
-    cm2.set(  h5 & 0xffffff); //3 history chars
-    cm2.set(  h5 ); //4 history chars
-
-    // large contexts (with hashtable)
-    //
-    U32 context, h6;
-
-    iCtxLarge.set( x.c4 >> 8, c1);
-    context = x.c4 & 0xffffff; //24 bits
-    h6 = iCtxLarge.get( context);
-    cm3.set(( h6 & 0xff)<<8+(context & 0xff));
-    cm3.set( ( h6 & 0xff) + (((32 * x.tt) & 0x1FFFFF00)));
-    cm3.set( h6<<8+ (context & 0xff)); //without context is also OK (not for news and obj2 however)
-
-    iCtxLarge.set( chars4 >> 8, c1);
-    context = chars4 & 0xffffff; //24 bits lowercase
-    h6 = iCtxLarge.get( context);
-    cm4.set((( h6 & 0xff)<<24)+  (context &0xffff));
-    cm4.set((h6 & 0xffff)+ context);
-    cm4.set( h6 );
-
-    iCtxLarge.set( buf(5) << 24 | x.c4 >> 8, c1);
-    context = x.c4; //32 bits
-    h6 = iCtxLarge.get( context);
-    cm5.set(( h6 & 0xff)<<16+  (context &0xffff));
-    cm5.set(( h6 & 0xffff)<<8+(context &0xff));
-    cm5.set(  h6 );
-  }
-  
-  cm2.mix(m);
-  cm3.mix(m);
-  cm4.mix(m);
-  cm5.mix(m);
-   cmt.mix(m);
-    cm0.mix(m);
-    cma.mix(m); 
-    cmc.mix(m); 
-  return 0;
-}
-virtual ~indirectModel1(){ }
-};
 
 //////////////////////////// dmcModel //////////////////////////
