+++ prt/contextmap.hpp
@@ -1,0 +1,113 @@
+#pragma once
+#include "types.hpp"
+#include "array.hpp"
+#include "mixer.hpp"
+#include "statemap.hpp"
+#include "rnd.hpp"
+#include "hash.hpp"
+
+// Context map for large contexts.  Most modeling uses this type of context
+// map.  It includes a built in RunContextMap to predict the last byte seen
+// in the same context, and also bit-level contexts that map to a bit
+// history state.
+//
+// Bit histories are stored in a hash table.  The table is organized into
+// 64-byte buckets alinged on cache page boundaries.  Each bucket contains
+// a hash chain of 7 elements, plus a 2 element queue (packed into 1 byte)
+// of the last 2 elements accessed for LRU replacement.  Each element has
+// a 2 byte checksum for detecting collisions, and an array of 7 bit history
+// states indexed by the last 0 to 2 bits of context.  The buckets are indexed
+// by a context ending after 0, 2, or 5 bits of the current byte.  Thus, each
+// byte modeled results in 3 main memory accesses per context, with all other
+// accesses to cache.
+//
+// On bits 0, 2 and 5, the context is updated and a new bucket is selected.
+// The most recently accessed element is tried first, by comparing the
+// 16 bit checksum, then the 7 elements are searched linearly.  If no match
+// is found, then the element with the lowest priority among the 5 elements
+// not in the LRU queue is replaced.  After a replacement, the queue is
+// emptied (so that consecutive misses favor a LFU replacement policy).
+// In all cases, the found/replaced element is put in the front of the queue.
+//
+// The priority is the state number of the first element (the one with 0
+// additional bits of context).  The states are sorted by increasing n0+n1
+// (number of bits seen), implementing a LFU replacement policy.
+//
+// When the context ends on a byte boundary (bit 0), only 3 of the 7 bit
+// history states are used.  The remaining 4 bytes implement a run model
+// as follows: <count:7,d:1> <b1> <unused> <unused> where <b1> is the last byte
+// seen, possibly repeated.  <count:7,d:1> is a 7 bit count and a 1 bit
+// flag (represented by count * 2 + d).  If d=0 then <count> = 1..127 is the
+// number of repeats of <b1> and no other bytes have been seen.  If d is 1 then
+// other byte values have been seen in this context prior to the last <count>
+// copies of <b1>.
+//
+// As an optimization, the last two hash elements of each byte (representing
+// contexts with 2-7 bits) are not updated until a context is seen for
+// a second time.  This is indicated by <count,d> = <1,0> (2).  After update,
+// <count,d> is updated to <2,0> or <1,1> (4 or 3).
+#if defined(__AVX2__)
+#define MALIGN 32
+#else
+#define MALIGN 16
+#endif
+
+
+class ContextMap {
+  const int C;  // max number of contexts
+  class E {  // hash element, 64 bytes
+    U16 chk[7];  // byte context checksums
+    U8 last;     // last 2 accesses (0-6) in low, high nibble
+  public:
+    U8 bh[7][7]; // byte context, 3-bit context -> bit history state
+      // bh[][0] = 1st bit, bh[][1,2] = 2nd bit, bh[][3..6] = 3rd bit
+      // bh[][0] is also a replacement priority, 0 = empty
+    U8* get(U16 ch)  // Find element (0-6) matching checksum.
+      // If not found, insert or replace lowest priority (not last).
+      {
+    
+  if (chk[last&15]==ch) return &bh[last&15][0];
+  int b=0xffff, bi=0;
+  for (int i=0; i<7; ++i) {
+    if (chk[i]==ch) return last=last<<4|i, (U8*)&bh[i][0];
+    int pri=bh[i][0];
+    if (pri<b && (last&15)!=i && last>>4!=i) b=pri, bi=i;
+  }
+  return last=0xf0|bi, chk[bi]=ch, (U8*)memset(&bh[bi][0], 0, 7);
+}
+  };
+  Array<E, 64> t;  // bit histories for bits 0-1, 2-4, 5-7
+    // For 0-1, also contains a run count in bh[][4] and value in bh[][5]
+    // and pending update count in bh[7]
+  Array<U8*> cp;   // C pointers to current bit history
+  Array<U8*> cp0;  // First element of 7 element array containing cp[i]
+  Array<U32> cxt;  // C whole byte contexts (hashes)
+  Array<U8*> runp; // C [0..3] = count, value, unused, unused
+  Array<short, MALIGN>  r0;   //for rle 
+  Array<short, MALIGN>  r1;
+  Array<short, MALIGN>  r0i;
+  Array<short, MALIGN>  rmask; // mask for skiped context
+  StateMap *sm;    // C maps of state -> p
+  int cn;          // Next context to set by set()
+  //void update(U32 cx, int c);  // train model that context cx predicts c
+  Random rnd;
+  int mix1(Mixer& m, int cc, int bp, int c1, int y1);
+   int mix2(Mixer& m, int s, StateMap& sm);
+    // mix() with global context passed as arguments to improve speed.
+  int model;  
+public:
+  ContextMap(U64 m, int c=1,int mod=-1);  // m = memory in bytes, a power of 2, C = c
+  ~ContextMap();
+  void set();   // set next whole byte context to cx
+  void set(U32 cx, int next=-1);   // set next whole byte context to cx
+  void set(U64 cx, int next=-1);   // set next whole byte context to cx
+    // if next is 0 then set order does not matter
+  //int mix(Mixer& m) { return mix1(m, m.x.c0, m.x.bpos, m.x.buf(1), m.x.y);}
+  int mix(Mixer& m) {return mix1(m, m.x.c0, m.x.bpos, m.x.buf(1), m.x.y);}
+    int inputs();
+};
+
+
+
+
+
